Store callstack index as uint32_t.

+++ server/TracyEvent.hpp
@@ -147,7 +147,7 @@ struct MemEvent
     uint64_t size;
     int64_t timeAlloc;
     int64_t timeFree;
-    uint64_t callstack;
+    uint32_t callstack;
     // All above is read/saved as-is.
 
     uint16_t threadAlloc;
